+++ src/Main.hs
@@ -69,7 +69,7 @@ main' = do
     -- FIXME: --all-epel and --all-releases
     <*> switchWith 'F' "all-fedora" "Query all Fedora releases"
     <*> (flagWith' [Source] 's' "source" "Query source repos" <|>
-         flagWith' allArchs 'A' "all-archs" "Query all (64 bit) arch repos" <|>
+         flagWith' allArchs 'A' "all-archs" "Query all arch repos" <|>
          many (optionWith (eitherReader eitherArch) 'a' "repo-arch" "ARCH" ("Specify repo arch [default:" +-+ showArch sysarch ++ "]")))
     <*> switchWith 't' "testing" "Fedora updates-testing"
     <*> switchWith 'D' "debug" "Show some debug output"