+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # Version history for fedora-repoquery
 
+## 0.9 (2026-06-26)
+- add --all-epel releases
+- c10s: use mirror.stream.centos.org
+- improve handling of archived epel major and minor releases
+- preliminary support for pre-stream centos
+- fedora updates Everything was introduced from F28
+- error for i386 arch unless epel6, f25, c7, or older
+- do not include i386 in --all-archs (regression in 0.8.2)
+
 ## 0.8.2 (2026-05-10)
 - `--all-archs`: also include source and i386
 - `--koji`: map epel10 to epel10.x